refactor: folder splash
