+++ lib/shared/utils/routes.dart
@@ -3,7 +3,7 @@ import '../../movements/view/movements_page.dart';
 import '../../portfolio/view/portfolio_page.dart';
 
 import '../../conversion/view/conversion_page.dart';
-import '../../splash_screen.dart';
+import '../../splash/splash_screen.dart';
 
 Map<String, Widget Function(BuildContext)> appRoutes = {
   SplashScreen.route: (context) => const SplashScreen(),
